ISSUE-20: Solve issue next/previous serie

+++ app/src/main/java/ar/com/tzulberti/archerytraining/model/tournament/TournamentSerieArrow.java
@@ -11,7 +11,7 @@ public class TournamentSerieArrow {
     public long id;
     public float xPosition;
     public float yPosition;
-    public long score;
+    public int score;
 
     public TournamentSerie serie;
     public Tournament tournament;
